feat: introduce modular task details components for enhanced management
- Added `TaskDetailsContainer`, `TaskDatesSection`, `TaskDescriptionSection`, and `TaskDetailsHeader` to streamline task management within the `TaskDetailsPage`.
- Each section encapsulates specific functionalities, such as editing titles, descriptions, and due dates, improving code organization and reusability.
- This modular approach enhances user experience by providing a clearer structure for task interactions, making it easier to manage task details effectively.
